remove token from ui code
