+++ src/authn/index.ts
@@ -1,7 +1,7 @@
 import { CREATE as POST } from 'data'
 import { clearApp } from 'data/storage'
 import { throwError } from '../error'
-import { clear as clearToken, getSeed, getToken } from './token'
+import { clearSeed, getSeed } from './seed'
 import { writable } from 'svelte/store'
 
 export type AuthLoginResponse = {
@@ -24,8 +24,8 @@ export const login = async (inviteCode = ''): Promise<void> => {
 }
 
 export const logout = async (): Promise<void> => {
-  const logoutUrl = `${process.env.API_HOST}/auth/logout?token=${encodeURIComponent(getToken())}`
+  const logoutUrl = `${process.env.API_HOST}/auth/logout}`
   clearApp()
-  clearToken()
+  clearSeed()
   location.replace(logoutUrl)
 }
