--- src/authn/token.ts
@@ -1,64 +1,0 @@
-import Cookies from 'js-cookie'
-
-const secure = process.env.CF_PAGES_BRANCH === 'main' || process.env.CF_PAGES_BRANCH === 'develop'
-
-export const getSeed = () => {
-  const seedCookie = Cookies.get('seed')
-  return seedCookie
-}
-
-export const getToken = () => Cookies.get('bearer-token') || ''
-
-export const clear = () => {
-  Cookies.remove('seed', { sameSite: 'strict', secure })
-  Cookies.remove('access-token', { sameSite: 'strict', secure })
-  Cookies.remove('token-type', { sameSite: 'strict', secure })
-  Cookies.remove('bearer-token', { sameSite: 'strict', secure })
-}
-
-initialize()
-
-function initialize() {
-  const seed = getSeed()
-  if (!seed) {
-    Cookies.set('seed', createSeed(), { expires: 7, sameSite: 'strict', secure })
-  }
-
-  initializeToken()
-}
-
-function initializeToken() {
-  const params = new URLSearchParams(location.search)
-
-  if (init('access-token') || init('token-type')) {
-    cleanAddressBar()
-  }
-
-  function init(name: string) {
-    const value = params.get(name)
-
-    if (value !== null) {
-      Cookies.set(name, value, { expires: 7, sameSite: 'strict', secure })
-      if (name === 'access-token') {
-        Cookies.set('bearer-token', `Bearer ${getSeed()}${value}`, { expires: 7, sameSite: 'strict', secure })
-        Cookies.remove('seed', { sameSite: 'strict', secure })
-        Cookies.remove('access-token', { sameSite: 'strict', secure })
-      }
-      params.delete(name)
-    }
-
-    return value
-  }
-
-  function cleanAddressBar() {
-    const search = [...params].length ? `?${params.toString()}` : ''
-
-    location.replace(location.pathname + search)
-  }
-}
-
-function createSeed() {
-  return Math.random()
-    .toString(36) // Convert to base-36 so we get more letters
-    .substring(2) // strip off the leading '0.'
-}
